Changing how loading is performed
