+++ project-gate/Assets/Scripts/PickUps/BaseGrenade.cs
@@ -1,0 +1,16 @@
+using Godot;
+
+namespace ProjGate.Pickups
+{
+  [GlobalClass]
+  public partial class BaseGrenade : Resource
+  {
+    public BaseGrenade()
+    {
+
+    }
+
+
+  }
+
+}
