+++ project-gate/Assets/Scripts/PickUps/BaseItem.cs
@@ -1,0 +1,12 @@
+using Godot;
+
+namespace ProjGate.Pickups
+{
+  [GlobalClass]
+  public class BaseItem : Resource
+  {
+    private int _index;
+    public string ItemName;
+    public Texture2D ItemIcon;
+  }
+}
