--- project-gate/Assets/Scripts/PickUps/Grenade.cs
@@ -1,10 +1,0 @@
-using Godot;
-
-public partial class Grenade : Resource {
-  public Grenade() {
-
-  }
-
-
-}
-
