+++ project-gate/Assets/Scripts/PickUps/PickupCatalogues/WeaponCatalogue.cs
@@ -1,0 +1,16 @@
+using Godot.Collections;
+using System.Collections.Generic;
+
+namespace ProjGate.Pickups
+{
+  public static class WeaponCatalogue
+  {
+    private static string WEAPON_CONFIGURATION_PATH => "res://Configuration";
+    public static List<BaseWeapon> PossibleWeapons;
+
+    /*public static async void GenerateWeaponCatalogue()
+    {
+    private const string CONFIGURATION_BASE_PATH = "res://Configuration/";
+  }*/
+  }
+}
