+++ project-gate/Assets/Scripts/PickUps/WeaponBehaviors/IEndTurnBehavior.cs
@@ -1,3 +1,7 @@
-public interface IEndTurnBehavior {
-  public void EndTurn();
+namespace ProjGate.Pickups
+{
+  public interface IEndTurnBehavior
+  {
+    public void EndTurn();
+  }
 }
