+++ project-gate/Assets/Scripts/PickUps/WeaponBehaviors/IGetTargetingBehavior.cs
@@ -1,3 +1,7 @@
-public interface IGetTargetingBehavior {
- public void GetTargetingInformation();
+namespace ProjGate.Pickups
+{
+  public interface IGetTargetingBehavior
+  {
+    public void GetTargetingInformation();
+  }
 }
