+++ project-gate/Assets/Units/playerteamcharacter.tscn
@@ -1,7 +1,8 @@
-[gd_scene load_steps=8 format=3 uid="uid://blamjf42tuppq"]
+[gd_scene load_steps=9 format=3 uid="uid://blamjf42tuppq"]
 
 [ext_resource type="Script" path="res://Assets/Scripts/Characters/PlayerTeamCharacter.cs" id="1_iabf2"]
 [ext_resource type="Material" uid="uid://k6f12crlx6sk" path="res://Assets/Materials/test_material.tres" id="2_43rut"]
+[ext_resource type="Resource" uid="uid://cbqdotqukeccr" path="res://Assets/Pickups/Weapons/test_weapon.tres" id="2_nsfu8"]
 [ext_resource type="AudioStream" uid="uid://d4ao63nnk7xvy" path="res://Assets/Sound-Effects/Dev-Only/Explosion3_modified.mp3" id="3_kpyht"]
 [ext_resource type="Script" path="res://Assets/Scripts/User-Interface/Healthbar.cs" id="4_wuko7"]
 [ext_resource type="Texture2D" uid="uid://bg6fhem324yn0" path="res://Assets/Textures/User-Interface/barHorizontal_green_mid 200.png" id="5_xpxo4"]
@@ -10,6 +11,7 @@
 
 [node name="playerteamcharacter" type="Node3D" groups=["PlayerCharacters"]]
 script = ExtResource("1_iabf2")
+MainWeapon = ExtResource("2_nsfu8")
 
 [node name="CSGCylinder3D" type="CSGCylinder3D" parent="."]
 material = ExtResource("2_43rut")
