+++ project-gate/Scenes/test_level.tscn
@@ -28,7 +28,10 @@ script = ExtResource("3_7h1qw")
 script = ExtResource("4_ht3ou")
 
 [node name="Level" type="Level" parent="."]
+m_level_should_generate_level = true
 m_level_file_location = "res://Configuration/Levels/Showrooms_1.yml"
+
+[node name="@TileGrid@17457" type="TileGrid" parent="Level"]
 
 [node name="Characters" type="Node" parent="." groups=["Characters"]]
